BE/requirements.txt
annotated-types==0.7.0
a2t==0.4.0
anyio==4.9.0
beautifulsoup4==4.13.4
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
curl_cffi==0.12.0
cycler==0.12.1
deprecation==2.1.0
fastapi==0.116.1
fonttools==4.59.0
frozendict==2.4.6
gotrue==2.12.3
greenlet==3.2.3
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
kiwisolver==1.4.8
matplotlib==3.10.3
multitasking==0.0.12
nltk==3.9.1
numpy==2.3.2
packaging==25.0
pandas==2.3.1
peewee==3.18.2
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
polars==1.31.0
postgrest==1.1.1
protobuf==6.31.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
realtime==2.6.0
requests==2.32.4
scikit-learn==1.6.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.47.2
StrEnum==0.4.15
torch==2.5.1
transformers==4.46.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
yfinance==0.2.65
